# Optimize `std::includes` by using `ranges::includes` approach