--- a/benchmarks/src/includes.cpp
+++ b/benchmarks/src/includes.cpp
@@ -3,9 +3,13 @@
 
 #include <algorithm>
 #include <benchmark/benchmark.h>
+#include <cstddef>
 #include <cstdint>
+#include <cstdlib>
 #include <iostream>
+#include <random>
 #include <ranges>
+#include <type_traits>
 #include <vector>
 
 #include "skewed_allocator.hpp"