# ❔ Why ranges approach

We have three outcomes of comparison of elements pair:
1. Haystack element less. Advance only haystack
2. Needle element less. We found mismatch, `return false`
3. Elements equal. Advance haystack and needle

Out of these outcomes, we can get one in one branch, the rest will be two branches.
Since we have "less" predicate, we can't have equal in one branch.

Out of _haystack less_ and _needle less_, the current `std::` algorithm favors _needle less_, and the ranges algorithm favors _haystack less_.
Favoring _haystack less_ optimizes long algorithm run, where haystack is way bigger than needle.
Favoring _needle less_ optimizes early return, which is not so useful, as it happens only once per algorithm run.

# 🏁 Benchmarks

Not sure what is common case of `includes`, trying these three:
(0) Needle is seen contiguously in haystack
(1) Needle is seen not completely contiguously in haystack, but still compact (stride is geometric distribution)
(2) Needle is spread in haystack with almost the same stride (plus-minus one)
(3) Needle is sampled randomly from haystack

All these are multiplied by
(1) The needle is actually found
(0) The middle element of the needle does not match anything in haystack

I think these combinations are enough to explore the algorithm performance properties

Benchmark results show being affected by codegen gremlin, and expectedly 300/290 cases are less faster and more slower.
But overall looks like an improvement.

# 📜 Benchmark results

⚠️ mixed result with improvements over 2.0 sometimes, but degradation below 0.3 sometimes too ⚠️ 

⚠️ whole benchmark runs long ⚠️ 

<details>
<summary>Mixed results with overall improvement, but some great degradation</summary>

Benchmark                                              |    Before   |     After    | Speedup
-------------------------------------------------------|-------------|--------------|---------
bm_includes<int8_t, alg_type::std_fn>/3000/3/0/1       |    771 ns   |    374 ns    |  2.06
bm_includes<int8_t, alg_type::std_fn>/3000/22/0/1      |    786 ns   |    453 ns    |  1.74
bm_includes<int8_t, alg_type::std_fn>/3000/105/0/1     |    809 ns   |    432 ns    |  1.87
bm_includes<int8_t, alg_type::std_fn>/3000/1504/0/1    |   1169 ns   |    924 ns    |  1.27
bm_includes<int8_t, alg_type::std_fn>/3000/2750/0/1    |   1363 ns   |   1362 ns    |  1.00
bm_includes<int8_t, alg_type::std_fn>/300/3/0/1        |   85.9 ns   |   57.7 ns    |  1.49
bm_includes<int8_t, alg_type::std_fn>/300/22/0/1       |   91.0 ns   |   76.9 ns    |  1.18
bm_includes<int8_t, alg_type::std_fn>/300/105/0/1      |    123 ns   |    112 ns    |  1.10
bm_includes<int8_t, alg_type::std_fn>/300/290/0/1      |    162 ns   |    155 ns    |  1.05
bm_includes<int8_t, alg_type::std_fn>/3000/3/0/0       |    733 ns   |    375 ns    |  1.95
bm_includes<int8_t, alg_type::std_fn>/3000/22/0/0      |    726 ns   |    392 ns    |  1.85
bm_includes<int8_t, alg_type::std_fn>/3000/105/0/0     |    720 ns   |    409 ns    |  1.76
bm_includes<int8_t, alg_type::std_fn>/3000/1504/0/0    |    729 ns   |    572 ns    |  1.27
bm_includes<int8_t, alg_type::std_fn>/3000/2750/0/0    |    732 ns   |    710 ns    |  1.03
bm_includes<int8_t, alg_type::std_fn>/300/3/0/0        |   86.1 ns   |   58.0 ns    |  1.48
bm_includes<int8_t, alg_type::std_fn>/300/22/0/0       |   86.1 ns   |   63.5 ns    |  1.36
bm_includes<int8_t, alg_type::std_fn>/300/105/0/0      |   84.6 ns   |   85.7 ns    |  0.99
bm_includes<int8_t, alg_type::std_fn>/300/290/0/0      |   83.0 ns   |   86.9 ns    |  0.96
bm_includes<int8_t, alg_type::std_fn>/3000/3/1/1       |    727 ns   |    376 ns    |  1.93
bm_includes<int8_t, alg_type::std_fn>/3000/22/1/1      |    826 ns   |    696 ns    |  1.19
bm_includes<int8_t, alg_type::std_fn>/3000/105/1/1     |    817 ns   |    780 ns    |  1.05
bm_includes<int8_t, alg_type::std_fn>/3000/1504/1/1    |   1098 ns   |   1042 ns    |  1.05
bm_includes<int8_t, alg_type::std_fn>/3000/2750/1/1    |   1401 ns   |   1355 ns    |  1.03
bm_includes<int8_t, alg_type::std_fn>/300/3/1/1        |   90.0 ns   |   73.0 ns    |  1.23
bm_includes<int8_t, alg_type::std_fn>/300/22/1/1       |    110 ns   |    114 ns    |  0.96
bm_includes<int8_t, alg_type::std_fn>/300/105/1/1      |    130 ns   |    153 ns    |  0.85
bm_includes<int8_t, alg_type::std_fn>/300/290/1/1      |    153 ns   |    154 ns    |  0.99
bm_includes<int8_t, alg_type::std_fn>/3000/3/1/0       |    729 ns   |    374 ns    |  1.95
bm_includes<int8_t, alg_type::std_fn>/3000/22/1/0      |    733 ns   |    385 ns    |  1.90
bm_includes<int8_t, alg_type::std_fn>/3000/105/1/0     |    756 ns   |    784 ns    |  0.96
bm_includes<int8_t, alg_type::std_fn>/3000/1504/1/0    |    762 ns   |    562 ns    |  1.36
bm_includes<int8_t, alg_type::std_fn>/3000/2750/1/0    |    744 ns   |    707 ns    |  1.05
bm_includes<int8_t, alg_type::std_fn>/300/3/1/0        |   88.1 ns   |   57.2 ns    |  1.54
bm_includes<int8_t, alg_type::std_fn>/300/22/1/0       |   90.7 ns   |   68.9 ns    |  1.32
bm_includes<int8_t, alg_type::std_fn>/300/105/1/0      |   93.8 ns   |    106 ns    |  0.88
bm_includes<int8_t, alg_type::std_fn>/300/290/1/0      |   92.7 ns   |   86.7 ns    |  1.07
bm_includes<int8_t, alg_type::std_fn>/3000/3/2/1       |   1228 ns   |    635 ns    |  1.93
bm_includes<int8_t, alg_type::std_fn>/3000/22/2/1      |   1554 ns   |   1010 ns    |  1.54
bm_includes<int8_t, alg_type::std_fn>/3000/105/2/1     |   2208 ns   |   1929 ns    |  1.14
bm_includes<int8_t, alg_type::std_fn>/3000/1504/2/1    |   2007 ns   |   2468 ns    |  0.81
bm_includes<int8_t, alg_type::std_fn>/3000/2750/2/1    |   1416 ns   |   1572 ns    |  0.90
bm_includes<int8_t, alg_type::std_fn>/300/3/2/1        |    151 ns   |    109 ns    |  1.39
bm_includes<int8_t, alg_type::std_fn>/300/22/2/1       |    199 ns   |    157 ns    |  1.27
bm_includes<int8_t, alg_type::std_fn>/300/105/2/1      |    230 ns   |    237 ns    |  0.97
bm_includes<int8_t, alg_type::std_fn>/300/290/2/1      |    143 ns   |    161 ns    |  0.89
bm_includes<int8_t, alg_type::std_fn>/3000/3/2/0       |    731 ns   |    393 ns    |  1.86
bm_includes<int8_t, alg_type::std_fn>/3000/22/2/0      |    863 ns   |    553 ns    |  1.56
bm_includes<int8_t, alg_type::std_fn>/3000/105/2/0     |   1079 ns   |    867 ns    |  1.24
bm_includes<int8_t, alg_type::std_fn>/3000/1504/2/0    |   1008 ns   |   1251 ns    |  0.81
bm_includes<int8_t, alg_type::std_fn>/3000/2750/2/0    |    699 ns   |    775 ns    |  0.90
bm_includes<int8_t, alg_type::std_fn>/300/3/2/0        |   96.3 ns   |   72.4 ns    |  1.33
bm_includes<int8_t, alg_type::std_fn>/300/22/2/0       |    108 ns   |   88.0 ns    |  1.23
bm_includes<int8_t, alg_type::std_fn>/300/105/2/0      |    122 ns   |    124 ns    |  0.98
bm_includes<int8_t, alg_type::std_fn>/300/290/2/0      |   72.0 ns   |   84.8 ns    |  0.85
bm_includes<int8_t, alg_type::std_fn>/3000/3/3/1       |   1277 ns   |    668 ns    |  1.91
bm_includes<int8_t, alg_type::std_fn>/3000/22/3/1      |   1542 ns   |   1080 ns    |  1.43
bm_includes<int8_t, alg_type::std_fn>/3000/105/3/1     |   2167 ns   |   1571 ns    |  1.38
bm_includes<int8_t, alg_type::std_fn>/3000/1504/3/1    |   2042 ns   |   2429 ns    |  0.84
bm_includes<int8_t, alg_type::std_fn>/3000/2750/3/1    |   1416 ns   |   2189 ns    |  0.65
bm_includes<int8_t, alg_type::std_fn>/300/3/3/1        |    169 ns   |    119 ns    |  1.42
bm_includes<int8_t, alg_type::std_fn>/300/22/3/1       |    206 ns   |    159 ns    |  1.30
bm_includes<int8_t, alg_type::std_fn>/300/105/3/1      |    241 ns   |    220 ns    |  1.10
bm_includes<int8_t, alg_type::std_fn>/300/290/3/1      |    142 ns   |    192 ns    |  0.74
bm_includes<int8_t, alg_type::std_fn>/3000/3/3/0       |    976 ns   |    521 ns    |  1.87
bm_includes<int8_t, alg_type::std_fn>/3000/22/3/0      |    977 ns   |    643 ns    |  1.52
bm_includes<int8_t, alg_type::std_fn>/3000/105/3/0     |   1082 ns   |    841 ns    |  1.29
bm_includes<int8_t, alg_type::std_fn>/3000/1504/3/0    |    990 ns   |   1170 ns    |  0.85
bm_includes<int8_t, alg_type::std_fn>/3000/2750/3/0    |    706 ns   |    782 ns    |  0.90
bm_includes<int8_t, alg_type::std_fn>/300/3/3/0        |    140 ns   |   98.1 ns    |  1.43
bm_includes<int8_t, alg_type::std_fn>/300/22/3/0       |   97.6 ns   |   87.5 ns    |  1.12
bm_includes<int8_t, alg_type::std_fn>/300/105/3/0      |    121 ns   |    112 ns    |  1.08
bm_includes<int8_t, alg_type::std_fn>/300/290/3/0      |   72.2 ns   |   93.9 ns    |  0.77
bm_includes<int16_t, alg_type::std_fn>/3000/3/0/1      |    724 ns   |    372 ns    |  1.95
bm_includes<int16_t, alg_type::std_fn>/3000/22/0/1     |    720 ns   |    414 ns    |  1.74
bm_includes<int16_t, alg_type::std_fn>/3000/105/0/1    |    746 ns   |    426 ns    |  1.75
bm_includes<int16_t, alg_type::std_fn>/3000/1504/0/1   |   1078 ns   |    918 ns    |  1.17
bm_includes<int16_t, alg_type::std_fn>/3000/2750/0/1   |   1371 ns   |   1369 ns    |  1.00
bm_includes<int16_t, alg_type::std_fn>/300/3/0/1       |   89.3 ns   |   54.7 ns    |  1.63
bm_includes<int16_t, alg_type::std_fn>/300/22/0/1      |   88.6 ns   |   79.3 ns    |  1.12
bm_includes<int16_t, alg_type::std_fn>/300/105/0/1     |    116 ns   |    108 ns    |  1.07
bm_includes<int16_t, alg_type::std_fn>/300/290/0/1     |    152 ns   |    157 ns    |  0.97
bm_includes<int16_t, alg_type::std_fn>/3000/3/0/0      |    753 ns   |    367 ns    |  2.05
bm_includes<int16_t, alg_type::std_fn>/3000/22/0/0     |    717 ns   |    378 ns    |  1.90
bm_includes<int16_t, alg_type::std_fn>/3000/105/0/0    |    928 ns   |    394 ns    |  2.36
bm_includes<int16_t, alg_type::std_fn>/3000/1504/0/0   |    728 ns   |    562 ns    |  1.30
bm_includes<int16_t, alg_type::std_fn>/3000/2750/0/0   |    729 ns   |    705 ns    |  1.03
bm_includes<int16_t, alg_type::std_fn>/300/3/0/0       |   83.1 ns   |   52.9 ns    |  1.57
bm_includes<int16_t, alg_type::std_fn>/300/22/0/0      |   85.5 ns   |   64.0 ns    |  1.34
bm_includes<int16_t, alg_type::std_fn>/300/105/0/0     |   89.3 ns   |   79.8 ns    |  1.12
bm_includes<int16_t, alg_type::std_fn>/300/290/0/0     |   82.7 ns   |   90.2 ns    |  0.92
bm_includes<int16_t, alg_type::std_fn>/3000/3/1/1      |    717 ns   |    370 ns    |  1.94
bm_includes<int16_t, alg_type::std_fn>/3000/22/1/1     |    734 ns   |    404 ns    |  1.82
bm_includes<int16_t, alg_type::std_fn>/3000/105/1/1    |    792 ns   |    532 ns    |  1.49
bm_includes<int16_t, alg_type::std_fn>/3000/1504/1/1   |   1077 ns   |    917 ns    |  1.17
bm_includes<int16_t, alg_type::std_fn>/3000/2750/1/1   |   1368 ns   |   1372 ns    |  1.00
bm_includes<int16_t, alg_type::std_fn>/300/3/1/1       |   87.2 ns   |   55.4 ns    |  1.57
bm_includes<int16_t, alg_type::std_fn>/300/22/1/1      |   97.7 ns   |   91.2 ns    |  1.07
bm_includes<int16_t, alg_type::std_fn>/300/105/1/1     |    136 ns   |    201 ns    |  0.68
bm_includes<int16_t, alg_type::std_fn>/300/290/1/1     |    151 ns   |    156 ns    |  0.97
bm_includes<int16_t, alg_type::std_fn>/3000/3/1/0      |    720 ns   |    368 ns    |  1.96
bm_includes<int16_t, alg_type::std_fn>/3000/22/1/0     |    718 ns   |    383 ns    |  1.87
bm_includes<int16_t, alg_type::std_fn>/3000/105/1/0    |    728 ns   |    451 ns    |  1.61
bm_includes<int16_t, alg_type::std_fn>/3000/1504/1/0   |    737 ns   |    561 ns    |  1.31
bm_includes<int16_t, alg_type::std_fn>/3000/2750/1/0   |    730 ns   |    703 ns    |  1.04
bm_includes<int16_t, alg_type::std_fn>/300/3/1/0       |   83.6 ns   |   52.2 ns    |  1.60
bm_includes<int16_t, alg_type::std_fn>/300/22/1/0      |   87.2 ns   |   66.3 ns    |  1.32
bm_includes<int16_t, alg_type::std_fn>/300/105/1/0     |   90.2 ns   |    130 ns    |  0.69
bm_includes<int16_t, alg_type::std_fn>/300/290/1/0     |   84.1 ns   |   90.4 ns    |  0.93
bm_includes<int16_t, alg_type::std_fn>/3000/3/2/1      |   1228 ns   |    643 ns    |  1.91
bm_includes<int16_t, alg_type::std_fn>/3000/22/2/1     |   1585 ns   |   1036 ns    |  1.53
bm_includes<int16_t, alg_type::std_fn>/3000/105/2/1    |   2156 ns   |   2358 ns    |  0.91
bm_includes<int16_t, alg_type::std_fn>/3000/1504/2/1   |   1818 ns   |   1905 ns    |  0.95
bm_includes<int16_t, alg_type::std_fn>/3000/2750/2/1   |   1536 ns   |   1458 ns    |  1.05
bm_includes<int16_t, alg_type::std_fn>/300/3/2/1       |    154 ns   |    110 ns    |  1.40
bm_includes<int16_t, alg_type::std_fn>/300/22/2/1      |    170 ns   |    189 ns    |  0.90
bm_includes<int16_t, alg_type::std_fn>/300/105/2/1     |    190 ns   |    239 ns    |  0.79
bm_includes<int16_t, alg_type::std_fn>/300/290/2/1     |    159 ns   |    230 ns    |  0.69
bm_includes<int16_t, alg_type::std_fn>/3000/3/2/0      |    746 ns   |    384 ns    |  1.94
bm_includes<int16_t, alg_type::std_fn>/3000/22/2/0     |    845 ns   |    554 ns    |  1.53
bm_includes<int16_t, alg_type::std_fn>/3000/105/2/0    |   1076 ns   |   1164 ns    |  0.92
bm_includes<int16_t, alg_type::std_fn>/3000/1504/2/0   |    922 ns   |   1093 ns    |  0.84
bm_includes<int16_t, alg_type::std_fn>/3000/2750/2/0   |    767 ns   |    757 ns    |  1.01
bm_includes<int16_t, alg_type::std_fn>/300/3/2/0       |   97.8 ns   |   68.8 ns    |  1.42
bm_includes<int16_t, alg_type::std_fn>/300/22/2/0      |   88.5 ns   |    107 ns    |  0.83
bm_includes<int16_t, alg_type::std_fn>/300/105/2/0     |    104 ns   |    125 ns    |  0.83
bm_includes<int16_t, alg_type::std_fn>/300/290/2/0     |   74.2 ns   |    120 ns    |  0.62
bm_includes<int16_t, alg_type::std_fn>/3000/3/3/1      |   1304 ns   |    672 ns    |  1.94
bm_includes<int16_t, alg_type::std_fn>/3000/22/3/1     |   1552 ns   |    990 ns    |  1.57
bm_includes<int16_t, alg_type::std_fn>/3000/105/3/1    |   2143 ns   |   1809 ns    |  1.18
bm_includes<int16_t, alg_type::std_fn>/3000/1504/3/1   |   2082 ns   |   3124 ns    |  0.67
bm_includes<int16_t, alg_type::std_fn>/3000/2750/3/1   |   1439 ns   |   1956 ns    |  0.74
bm_includes<int16_t, alg_type::std_fn>/300/3/3/1       |    180 ns   |    123 ns    |  1.46
bm_includes<int16_t, alg_type::std_fn>/300/22/3/1      |    181 ns   |    220 ns    |  0.82
bm_includes<int16_t, alg_type::std_fn>/300/105/3/1     |    192 ns   |    278 ns    |  0.69
bm_includes<int16_t, alg_type::std_fn>/300/290/3/1     |    145 ns   |    215 ns    |  0.67
bm_includes<int16_t, alg_type::std_fn>/3000/3/3/0      |    999 ns   |    515 ns    |  1.94
bm_includes<int16_t, alg_type::std_fn>/3000/22/3/0     |    991 ns   |    623 ns    |  1.59
bm_includes<int16_t, alg_type::std_fn>/3000/105/3/0    |   1077 ns   |    886 ns    |  1.22
bm_includes<int16_t, alg_type::std_fn>/3000/1504/3/0   |    991 ns   |   1540 ns    |  0.64
bm_includes<int16_t, alg_type::std_fn>/3000/2750/3/0   |    719 ns   |   1058 ns    |  0.68
bm_includes<int16_t, alg_type::std_fn>/300/3/3/0       |    144 ns   |   91.1 ns    |  1.58
bm_includes<int16_t, alg_type::std_fn>/300/22/3/0      |   83.7 ns   |    118 ns    |  0.71
bm_includes<int16_t, alg_type::std_fn>/300/105/3/0     |   94.4 ns   |    136 ns    |  0.69
bm_includes<int16_t, alg_type::std_fn>/300/290/3/0     |   73.5 ns   |    106 ns    |  0.69
bm_includes<int32_t, alg_type::std_fn>/3000/3/0/1      |    728 ns   |    386 ns    |  1.89
bm_includes<int32_t, alg_type::std_fn>/3000/22/0/1     |    742 ns   |    387 ns    |  1.92
bm_includes<int32_t, alg_type::std_fn>/3000/105/0/1    |    764 ns   |    424 ns    |  1.80
bm_includes<int32_t, alg_type::std_fn>/3000/1504/0/1   |   1101 ns   |    916 ns    |  1.20
bm_includes<int32_t, alg_type::std_fn>/3000/2750/0/1   |   1401 ns   |   1345 ns    |  1.04
bm_includes<int32_t, alg_type::std_fn>/300/3/0/1       |   86.5 ns   |   58.2 ns    |  1.49
bm_includes<int32_t, alg_type::std_fn>/300/22/0/1      |   91.7 ns   |   72.1 ns    |  1.27
bm_includes<int32_t, alg_type::std_fn>/300/105/0/1     |    123 ns   |    105 ns    |  1.17
bm_includes<int32_t, alg_type::std_fn>/300/290/0/1     |    152 ns   |    152 ns    |  1.00
bm_includes<int32_t, alg_type::std_fn>/3000/3/0/0      |    733 ns   |    375 ns    |  1.95
bm_includes<int32_t, alg_type::std_fn>/3000/22/0/0     |    729 ns   |    385 ns    |  1.89
bm_includes<int32_t, alg_type::std_fn>/3000/105/0/0    |    746 ns   |    401 ns    |  1.86
bm_includes<int32_t, alg_type::std_fn>/3000/1504/0/0   |    737 ns   |    559 ns    |  1.32
bm_includes<int32_t, alg_type::std_fn>/3000/2750/0/0   |    742 ns   |    706 ns    |  1.05
bm_includes<int32_t, alg_type::std_fn>/300/3/0/0       |   88.9 ns   |   57.5 ns    |  1.55
bm_includes<int32_t, alg_type::std_fn>/300/22/0/0      |   86.8 ns   |   70.4 ns    |  1.23
bm_includes<int32_t, alg_type::std_fn>/300/105/0/0     |   94.4 ns   |   81.0 ns    |  1.17
bm_includes<int32_t, alg_type::std_fn>/300/290/0/0     |   83.3 ns   |   83.2 ns    |  1.00
bm_includes<int32_t, alg_type::std_fn>/3000/3/1/1      |    730 ns   |    374 ns    |  1.95
bm_includes<int32_t, alg_type::std_fn>/3000/22/1/1     |    754 ns   |    405 ns    |  1.86
bm_includes<int32_t, alg_type::std_fn>/3000/105/1/1    |    832 ns   |    512 ns    |  1.63
bm_includes<int32_t, alg_type::std_fn>/3000/1504/1/1   |   1104 ns   |    912 ns    |  1.21
bm_includes<int32_t, alg_type::std_fn>/3000/2750/1/1   |   1393 ns   |   1343 ns    |  1.04
bm_includes<int32_t, alg_type::std_fn>/300/3/1/1       |   87.0 ns   |   57.1 ns    |  1.52
bm_includes<int32_t, alg_type::std_fn>/300/22/1/1      |    108 ns   |   87.5 ns    |  1.23
bm_includes<int32_t, alg_type::std_fn>/300/105/1/1     |    129 ns   |    156 ns    |  0.83
bm_includes<int32_t, alg_type::std_fn>/300/290/1/1     |    153 ns   |    151 ns    |  1.01
bm_includes<int32_t, alg_type::std_fn>/3000/3/1/0      |    727 ns   |    372 ns    |  1.95
bm_includes<int32_t, alg_type::std_fn>/3000/22/1/0     |    733 ns   |    385 ns    |  1.90
bm_includes<int32_t, alg_type::std_fn>/3000/105/1/0    |    762 ns   |    442 ns    |  1.72
bm_includes<int32_t, alg_type::std_fn>/3000/1504/1/0   |    741 ns   |    565 ns    |  1.31
bm_includes<int32_t, alg_type::std_fn>/3000/2750/1/0   |    758 ns   |    708 ns    |  1.07
bm_includes<int32_t, alg_type::std_fn>/300/3/1/0       |   86.4 ns   |   56.6 ns    |  1.53
bm_includes<int32_t, alg_type::std_fn>/300/22/1/0      |   92.0 ns   |   71.2 ns    |  1.29
bm_includes<int32_t, alg_type::std_fn>/300/105/1/0     |   95.9 ns   |    107 ns    |  0.90
bm_includes<int32_t, alg_type::std_fn>/300/290/1/0     |   81.5 ns   |   83.8 ns    |  0.97
bm_includes<int32_t, alg_type::std_fn>/3000/3/2/1      |   1200 ns   |    635 ns    |  1.89
bm_includes<int32_t, alg_type::std_fn>/3000/22/2/1     |   1577 ns   |   1030 ns    |  1.53
bm_includes<int32_t, alg_type::std_fn>/3000/105/2/1    |   2212 ns   |   1654 ns    |  1.34
bm_includes<int32_t, alg_type::std_fn>/3000/1504/2/1   |   1499 ns   |   1211 ns    |  1.24
bm_includes<int32_t, alg_type::std_fn>/3000/2750/2/1   |   1413 ns   |   1365 ns    |  1.04
bm_includes<int32_t, alg_type::std_fn>/300/3/2/1       |    148 ns   |    107 ns    |  1.38
bm_includes<int32_t, alg_type::std_fn>/300/22/2/1      |    213 ns   |    170 ns    |  1.25
bm_includes<int32_t, alg_type::std_fn>/300/105/2/1     |    194 ns   |    215 ns    |  0.90
bm_includes<int32_t, alg_type::std_fn>/300/290/2/1     |    166 ns   |    225 ns    |  0.74
bm_includes<int32_t, alg_type::std_fn>/3000/3/2/0      |    738 ns   |    392 ns    |  1.88
bm_includes<int32_t, alg_type::std_fn>/3000/22/2/0     |    857 ns   |    584 ns    |  1.47
bm_includes<int32_t, alg_type::std_fn>/3000/105/2/0    |   1142 ns   |    846 ns    |  1.35
bm_includes<int32_t, alg_type::std_fn>/3000/1504/2/0   |    778 ns   |    606 ns    |  1.28
…