@AlexGuteniev 2025-06-03
I think the chance of UB were lower, than your commit message says. We pick the middle needle elemnt to break the match, not an arbitrary random number.

Otherwise fine.

Please confirm that my results  where some cases became worse, don't concern you too much.