@StephanTLavavej 2025-06-03
> I think the chance of UB were lower, than your commit message says.

Good point, yeah. :game_die:

> Please confirm that my results where some cases became worse, don't concern you too much.

Yeah, that's fine. Thanks for the analysis! :smile_cat: